+++ CHANGELOG.md
@@ -288,3 +288,6 @@ v2.5.1 (2024-04-02)
 - **ticloud** module:
   - Added the following text to the docstrings for the `ticloud.URLThreatIntelligence.get_url_analysis_feed_from_date` and `ticloud.URLThreatIntelligence.get_url_analysis_feed_from_date_aggregated` methods: "It is possible to list analyses up to 90 days into the past."
   - Added the `get_objects_aggregated` method to the `TAXIIRansomwareFeed` class.
+
+- **a1000** module:
+  - The `a1000.A1000.get_yara_ruleset_matches_v2` method now also accepts a list u of multiple ruleset names as the `ruleset_name` parameter.